JGraphs

Documentation for JGraphs.

DO NOT USE YET

JGraphs.JGraph

JGraphs.JGraphData

JGraphs.GB2Luxor

JGraphs.draw_node

JGraphs.jedges

JGraphs.jgraph_morph

JGraphs.jgraph_walk

JGraphs.jnodes

JGraphs.JGraph — Type

JGraph(g::JGraphData)

The JGraph structure is the core structure of the package when called with a JGraphData argument creates and Object for each node in the graph and stores them in jnodes and an Object for each edge and stores them in jedges. Note that the nodes are stored in a Vector, whereas the edges are stored in a Dict with pairs of nodes (integers) as keys.

***Note*** Since JGraph stores Objects it can only be called after a Video has been initialized look at Javis documentation to know all about it.

JGraphs.JGraphData — Type

JGraphData(graph::AbstractGraph, layout::NetworkLayout.AbstractLayout; kwargs...)

The JGraphData is the structure used to set all the properties of the graph Animation, the graph itself, its layout and all the properties useful to animate it in the appropriate way. It should be passed to JGraph to obtain an animation with the specified properties. The keyword arguments are supposed to increase in the future.

Arguments

graph::AbstractGraph: A graph as created with LightGraphs and affiliate packages

layout::NetworkLayout.AbstractLayout: A layout as provided by NetworkLayout to place the

graph nodes in sapce.

Keywords

node_color sets the color of all nodes

node_size sets the size of all nodes

edges_color sets the colors of all edges

edges_width sets the width of all edges

scaling scales the positions of all nodes

frames the frames during which the animations on the graph will last if not specified elsewhere

numbered Defaults to false if set to true labels the nodes with their number as vertices

in the graph.

JGraphs.GB2Luxor — Method

GB2Luxor(p)

Transforms a Point defined in GeometryBasics into one defined in Luxor.

JGraphs.draw_node — Method

draw_node(args; center, radius, action)

Function used to draw a circle. Very similar to Javis.JCircle with the only difference that it does not use sethue so animating a color change is easier.

JGraphs.jedges — Method

jedges(jg::JGraph)

Accessor function to get access to the jedges stored in a JGraph.

JGraphs.jgraph_morph — Method

jgraph_morph(g::JGraph, layout_to, scaling_to; frames=nothing)

This functions allows to animate the change from one layout to another for a graph.

Arguments

g::JGraph the JGraph to animate.

layout_to a NetworkLayout.AbstractLayout that will be applied to nodes and edges

to determine the configuration to reach.

scaling_to an scaling factor to scale the positions of nodes and edges in the animation.

Keywords

frames a range of frames during which the the morph will take place.

If not specified the animation will last as long aas the JGraph itself.

JGraphs.jgraph_walk — Method

jgraph_walk(g::JGraph, walk_func)

This function will iteratively change the colors of nodes and edges in a graph to draw a path along them of the graph.

Arguments

g::JGraph the JGraph on which the walk should be animated

walk_func a function that takes as argument and AbstractGraph and returns

a set of nodes linked each to the following one by and edge present in the graph. One can use randomwalk and non_backtracking_randomwalk from LightGraphs.

JGraphs.jnodes — Method

jnodes(jg::JGraph)

Accessor function to get access to the jnodes stored in a JGraph.